+++ app/Http/Parameters/Get.php
@@ -176,7 +176,7 @@ class Get
                     if (array_key_exists($key, self::$parameters) === true) {
                         if (
                             is_string(self::$parameters[$key]) === false ||
-                            in_array(self::$parameters[$key], ['api', 'legacy', 'website']) === false
+                            in_array(self::$parameters[$key], ['api', 'legacy', 'postman', 'website']) === false
                         ) {
                             unset(self::$parameters[$key]);
                         }